requirements.txt
fastapi
supabase
jwt
uvicorn
httpx
python-dotenv